perf: isolate public edge url resolution
